Split utilities into 2 header files and update the code for better readability
