+++ src/App/RayTracerLayer.cpp
@@ -11,19 +11,20 @@
 #include <cstdint>
 #include <vector>
 
-#include "Core/Color.hpp"
-#include "Core/Ray.hpp"
-#include "Core/Sphere.hpp"
-
-#include "Utils/Timer.hpp"
-#include "Utils/Random.hpp"
-#include "Utils/Utils.hpp"
-
 #include "Renderer/VertexArray.hpp"
 #include "Renderer/Buffer.hpp"
 #include "Renderer/Image.hpp"
 #include "Renderer/Shader.hpp"
 #include "Renderer/Renderer.hpp"
+
+#include "Utils/Timer.hpp"
+#include "Utils/Random.hpp"
+#include "Utils/RayTracerUtils.hpp"
+#include "Utils/GLUtils.hpp"
+
+#include "Core/Color.hpp"
+#include "Core/Ray.hpp"
+#include "Core/Sphere.hpp"
 
 namespace fs = std::filesystem;
 
@@ -65,8 +66,8 @@ namespace raytracer {
         m_VertexArray->AddVertexBuffer(vertexBuffer, 0, 3, GL_FLOAT, GL_FALSE, 5 * sizeof(float), (void*)0);
         m_VertexArray->AddVertexBuffer(vertexBuffer, 1, 2, GL_FLOAT, GL_FALSE, 5 * sizeof(float), (void*)(3 * sizeof(float)));
 
-
-                m_World.Add(MakeRef<Sphere>(glm::vec3{0, 0, -1}, 0.5));
+        // Add objects to the world
+        m_World.Add(MakeRef<Sphere>(glm::vec3{0, 0, -1}, 0.5));
         m_World.Add(MakeRef<Sphere>(glm::vec3{0, -100.5, -1}, 100));
     }
