+++ src/App/RayTracerLayer.hpp
@@ -13,6 +13,8 @@
 #include "Core/Color.hpp"
 #include "Core/HittableList.hpp"
 
+#include "Utils/RayTracerUtils.hpp"
+
 namespace raytracer {
 
     class RayTracerLayer : public Layer
@@ -33,10 +35,10 @@ namespace raytracer {
         std::uint32_t m_ViewportWidth = 0, m_ViewportHeight = 0;
         float m_LastRenderTime = 0.0;
 
-        std::shared_ptr<VertexArray> m_VertexArray;
-        std::shared_ptr<Shader> m_Shader;
+        Ref<VertexArray> m_VertexArray;
+        Ref<Shader> m_Shader;
 
-        std::shared_ptr<Image> m_Image;
+        Ref<Image> m_Image;
         std::vector<std::uint8_t> m_ImageData;
 
         HittableList m_World;
