+++ src/App/Window.cpp
@@ -6,7 +6,8 @@
 #include "Renderer/OpenGLHeaders.hpp"
 
 #include "Utils/Result.hpp"
-#include "Utils/Utils.hpp"
+#include "Utils/RayTracerUtils.hpp"
+#include "Utils/GLUtils.hpp"
 
 namespace raytracer {
