+++ src/Renderer/Buffer.cpp
@@ -2,7 +2,8 @@
 
 #include "OpenGLHeaders.hpp"
 
-#include "Utils/Utils.hpp"
+#include "Utils/RayTracerUtils.hpp"
+#include "Utils/GLUtils.hpp"
 
 namespace raytracer {
