+++ src/Renderer/Shader.hpp
@@ -8,7 +8,7 @@
 #include "OpenGLHeaders.hpp"
 
 #include "Utils/Result.hpp"
-#include "Utils/Utils.hpp"
+#include "Utils/RayTracerUtils.hpp"
 
 namespace raytracer {
