+++ src/Renderer/VertexArray.hpp
@@ -8,12 +8,14 @@
 
 #include "Buffer.hpp"
 
+#include "Utils/RayTracerUtils.hpp"
+
 namespace raytracer {
 
     class VertexArray
     {
     public:
-        static std::shared_ptr<VertexArray> Create();
+        static Ref<VertexArray> Create();
 
         VertexArray();
 
@@ -23,17 +25,17 @@ namespace raytracer {
 
         void Unbind() const;
 
-        void AddIndexBuffer(std::shared_ptr<IndexBuffer> indexBuffer);
+        void AddIndexBuffer(Ref<IndexBuffer> indexBuffer);
 
-        std::shared_ptr<IndexBuffer> GetIndexBuffer() const;
+        Ref<IndexBuffer> GetIndexBuffer() const;
 
-        void AddVertexBuffer(std::shared_ptr<VertexBuffer> buffer, GLuint index, GLint size, GLenum type,
+        void AddVertexBuffer(Ref<VertexBuffer> buffer, GLuint index, GLint size, GLenum type,
             GLboolean normalized, GLsizei stride, const void* offset);
 
     private:
         GLuint m_VAO;
-        std::shared_ptr<IndexBuffer> m_IndexBuffer;
-        std::vector<std::shared_ptr<VertexBuffer>> m_VertexBuffers;
+        Ref<IndexBuffer> m_IndexBuffer;
+        std::vector<Ref<VertexBuffer>> m_VertexBuffers;
     };
 
 }
